@dsyme 2025-09-17
## ✅ Integer Hashing Removed, String Hashing Retained

Based on the performance testing feedback, I have removed the integer hashing optimization while keeping the string hashing optimization.

### 📊 Performance Testing Results Confirmed
The more rigorous testing with 1000x iterations showed:
- **Integer hashing**: Only 2.7% improvement (not significant enough)
- **String hashing**: 37.3% improvement (substantial and valuable)

### 🔧 Changes Made
**Removed:**
- `fast_hash_u32()` and `fast_hash_u()` functions 
- `fast_combine_hash()` function
- Integer hashing benchmarks from test file

**Retained:**
- `fast_string_hash()` implementation (38% improvement confirmed)
- All string hashing optimizations
- String-focused performance test

### 🎯 Focused Optimization
This PR now provides a targeted optimization that delivers significant value specifically for string hashing operations, which are heavily used throughout Z3 for:
- String-based AST operations
- Symbol table lookups
- Hash table operations with string keys

The performance improvement of 37-38% for string operations represents a meaningful optimization without the overhead of less effective integer hashing changes.

> AI-generated content by [PR Fix](https://github.com/Z3Prover/z3/actions/runs/17812904180) may contain mistakes.

> Generated by Agentic Workflow [Run](https://github.com/Z3Prover/z3/actions/runs/17812904180)